+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 All notable changes to the "wasm" extension will be documented in this file.
 
 ## [Unreleased]
-- Added icon (@bitjson)
+- Added icon ([@bitjson](https://github.com/bitjson))
+- Support large wasm files. Now a pure js wasm parser / printer [webassemblyjs] is used.(https://github.com/xtuc/webassemblyjs).
 
 ## [1.0.1] - 2017-12-04
 - Fix badges only.